# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
            "api.ChangePasswordRequest": {
                "properties": {
                    "current_password": {
                        "type": "string"
                    },
                    "new_password": {
                        "type": "string"
                    }
                },
                "required": [
                    "current_password",
                    "new_password"
                ],
                "type": "object"
            },
            "api.CreateAPIKeyRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
            "api.CreateAPIKeyResponse": {
                "properties": {
                    "api_key": {
                        "$ref": "#/components/schemas/types.APIKey"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "api_key",
                    "token"
                ],
                "type": "object"
            },
...
            "api.CurrentUser": {
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                },
                "required": [
                    "role",
                    "username"
                ],
                "type": "object"
            },
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "api.LoginResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ],
                "type": "object"
            },
...
            "api.StatusResponse": {
                "properties": {
                    "status": {
                        "type": "string"
                    }
                },
                "required": [
                    "status"
                ],
                "type": "object"
            },
...
            "types.APIKey": {
                "properties": {
                    "created_at": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "string"
                    },
                    "last_used_at": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "prefix": {
                        "type": "string"
                    },
                    "revoked_at": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "created_at",
                    "id",
                    "last_used_at",
                    "name",
                    "prefix",
                    "revoked_at",
                    "user_id"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/me": {
            "get": {
                "operationId": "currentUser",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.CurrentUser"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "currentUser",
                "tags": [
                    "login"
                ]
            }
        },
        "/api/me/api-keys": {
            "get": {
                "operationId": "listAPIKeys",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/types.APIKey"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listAPIKeys",
                "tags": [
                    "api-keys"
                ]
            },
            "post": {
                "operationId": "createAPIKey",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/api.CreateAPIKeyRequest",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.CreateAPIKeyResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createAPIKey",
                "tags": [
                    "api-keys"
                ]
            }
        },
        "/api/me/api-keys/{id}": {
            "delete": {
                "operationId": "revokeAPIKey",
                "parameters": [
                    {
                        "description": "API key id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.StatusResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "revokeAPIKey",
                "tags": [
                    "api-keys"
                ]
            }
        },
        "/api/me/password": {
            "patch": {
                "operationId": "changePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/api.ChangePasswordRequest",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.LoginResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "changePassword",
                "tags": [
                    "login"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}